Add injected function for loading the ROOT tree of nnL simulation
Similar to the function responsible for the H3L and H4L case, it
loads the events of ROOT tree from the GEANT4 simulation of nnL
of the solenoid setup. The difference is that the charge of the
decay fragment is 1 in this case compared to the case of H3L or
H4L (charge is 2 in this case).